move BatchDriftGenerator to public module
